# openapi.yaml
openapi: 3.0.3
info:
  title: J1Hub API
  version: 1.0.0
  description: |
    Stable API contract for J1Hub endpoints.

    ## Versioning plan
    - Canonical versioned base path: `/api/v1`.
    - New clients use `/api/v1/*`; unversioned routes remain available while `API_ENABLE_LEGACY_ROUTES=true`.
    - New clients should migrate to `/api/v1/*` immediately.

    ## Transition strategy
    1. **Now**: support both `/api/v1/*` and legacy endpoints.
    2. **Deprecation window**: communicate deprecation for legacy routes via release notes and docs.
    3. **Removal**: disable legacy aliases by configuration after migration completion.
servers:
  - url: /
    description: Legacy (transition period)
  - url: /api/v1
    description: Canonical versioned API
security:
  - bearerAuth: []
paths:
...
  /billing/status:
    get:
      summary: Get employer billing entitlements
      operationId: getBillingStatus
      responses:
        '200':
          description: Current credit and subscription status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingStatus'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /billing/history:
    get:
      summary: List recent employer billing events
      operationId: getBillingHistory
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Recent billing events
          content:
            application/json:
              schema:
                type: object
                required: [events, count]
                properties:
                  events:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /billing/create-checkout-session:
    post:
      summary: Create Stripe Checkout session
      operationId: createCheckoutSession
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                purchase_type:
                  type: string
                  enum: [listing, subscription]
                quantity:
                  type: integer
                success_url:
                  type: string
                cancel_url:
                  type: string
      responses:
        '200':
          description: Checkout session
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, url]
                properties:
                  sessionId:
                    type: string
                  url:
                    type: string
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
  /billing/webhook:
    post:
      summary: Stripe webhook
      operationId: billingWebhook
      security: []
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [success, duplicate]
                    example: success
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: JSON error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error, detail, request_id]
      properties:
        error:
          type: string
        detail:
          type: string
        request_id:
          type: string
...
    BillingStatus:
      type: object
      required: [listing_credits, has_active_subscription, active_until]
      properties:
        listing_credits:
          type: integer
        has_active_subscription:
          type: boolean
        active_until:
          type: string
          format: date-time
          nullable: true
...